Compress detection data modules to one file
