--- src/netspresso_trainer/dataloaders/detection/__init__.py
@@ -1,2 +1,0 @@
-from .dataset import DetectionSampleLoader
-from .local import DetectionCustomDataset
